+++ tests/evaluators/test_code_block_writer.py
@@ -14,6 +14,7 @@ from sybil_extras.languages import (
     MARKDOWN_IT,
     MDX,
     MYST,
+    MYST_PARSER,
     NORG,
     RESTRUCTUREDTEXT,
     MarkupLanguage,
@@ -69,6 +70,7 @@ def test_writes_modified_content(
         DJOT: markdown_content,
         NORG: norg_content,
         MYST: myst_content,
+        MYST_PARSER: markdown_content,
     }[markup_language]
 
     source_file = tmp_path / "source_file.txt"
@@ -131,6 +133,7 @@ def test_writes_modified_content(
         DJOT: markdown_expected,
         NORG: norg_expected,
         MYST: myst_expected,
+        MYST_PARSER: markdown_expected,
     }[markup_language]
 
     assert source_file.read_text(encoding="utf-8") == expected_content
@@ -248,6 +251,7 @@ def test_empty_code_block_write_content(
         DJOT: markdown_content,
         NORG: norg_content,
         MYST: myst_content,
+        MYST_PARSER: markdown_content,
     }[markup_language]
     source_file = tmp_path / "source_file.txt"
     source_file.write_text(data=content, encoding="utf-8")
@@ -324,6 +328,7 @@ def test_empty_code_block_write_content(
         DJOT: markdown_expected,
         NORG: norg_expected,
         MYST: myst_expected,
+        MYST_PARSER: markdown_expected,
     }[markup_language]
 
     example.evaluate()
@@ -340,7 +345,7 @@ def test_empty_code_block_with_options(
     code
     block has options.
     """
-    if markup_language in (MARKDOWN, MARKDOWN_IT, DJOT, NORG):
+    if markup_language in (MARKDOWN, MARKDOWN_IT, MYST_PARSER, DJOT, NORG):
         # Markdown-like formats do not support code block options.
         return
 
@@ -753,6 +758,7 @@ def test_indented_existing_block(
         DJOT: markdown_content,
         NORG: norg_content,
         MYST: myst_content,
+        MYST_PARSER: markdown_content,
     }[markup_language]
     source_file = tmp_path / "source_file.txt"
     source_file.write_text(data=original_content, encoding="utf-8")
@@ -816,6 +822,7 @@ def test_indented_existing_block(
         DJOT: markdown_expected,
         NORG: norg_expected,
         MYST: myst_expected,
+        MYST_PARSER: markdown_expected,
     }[markup_language]
     assert source_file_content == expected_content
 
@@ -876,6 +883,7 @@ def test_indented_empty_existing_block(
         DJOT: markdown_content,
         NORG: norg_content,
         MYST: myst_content,
+        MYST_PARSER: markdown_content,
     }[markup_language]
     source_file = tmp_path / "source_file.txt"
     source_file.write_text(data=original_content, encoding="utf-8")
@@ -947,6 +955,7 @@ def test_indented_empty_existing_block(
         DJOT: markdown_expected,
         NORG: norg_expected,
         MYST: myst_expected,
+        MYST_PARSER: markdown_expected,
     }[markup_language]
     assert source_file_content == expected_content