+++ tests/evaluators/test_shell_evaluator.py
@@ -36,6 +36,7 @@ from sybil_extras.languages import (
     MARKDOWN_IT,
     MDX,
     MYST,
+    MYST_PARSER,
     NORG,
     RESTRUCTUREDTEXT,
     MarkupLanguage,
@@ -443,6 +444,7 @@ def test_write_to_file_new_content_trailing_newlines(
         DJOT: markdown_content,
         NORG: norg_content,
         MYST: myst_content,
+        MYST_PARSER: markdown_content,
     }[markup_language]
     source_file = tmp_path / "source_file.txt"
     source_file.write_text(data=original_content, encoding="utf-8")
@@ -517,6 +519,7 @@ def test_write_to_file_new_content_trailing_newlines(
         DJOT: markdown_expected,
         NORG: norg_expected,
         MYST: myst_expected,
+        MYST_PARSER: markdown_expected,
     }[markup_language]
     if write_to_file:
         assert source_file_content == expected_content
@@ -585,6 +588,7 @@ def test_write_to_file_new_content_no_trailing_newlines(
         DJOT: markdown_content,
         NORG: norg_content,
         MYST: myst_content,
+        MYST_PARSER: markdown_content,
     }[markup_language]
     source_file = tmp_path / "source_file.txt"
     source_file.write_text(data=original_content, encoding="utf-8")
@@ -654,6 +658,7 @@ def test_write_to_file_new_content_no_trailing_newlines(
         DJOT: markdown_expected,
         NORG: norg_expected,
         MYST: myst_expected,
+        MYST_PARSER: markdown_expected,
     }[markup_language]
     if write_to_file:
         assert source_file_content == expected_content